+++ C# and Project/Program.cs
@@ -1,3 +1,5 @@
+using C__and_Project.Repositories;
+
 namespace C__and_Project
 {
     public class Program
@@ -6,15 +8,18 @@ namespace C__and_Project
         {
             var builder = WebApplication.CreateBuilder(args);
 
-            // Add services to the container.
             builder.Services.AddControllersWithViews();
+
+            //builder.Services.AddSingleton<IUsersRepository, DummyUsersRepository>();
+
+            builder.Services.AddSingleton<IUsersRepository, UsersRepository>();
 
             var app = builder.Build();
 
             // Configure the HTTP request pipeline.
             if (!app.Environment.IsDevelopment())
             {
-                app.UseExceptionHandler("/Home/zero");
+                app.UseExceptionHandler("/Home/Error");
                 // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
                 app.UseHsts();
             }
@@ -28,11 +33,9 @@ namespace C__and_Project
 
             app.MapControllerRoute(
                 name: "default",
-                pattern: "{controller=Home}/{action=Index}/{id?}");
+                pattern: "{controller=Users}/{action=Index}/{id?}");
 
             app.Run();
         }
     }
 }
-
-//help me plase
