+++ C# and Project/appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "C#andProjectDatabase": "Server=tcp:projectserver-01.database.windows.net,1433;Initial Catalog=Project 1B Project;Persist Security Info=False;User ID=admin012;Password={your_password};MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;"
+    "C#andProjectDatabase": "Server=tcp:projectserver-01.database.windows.net,1433;Initial Catalog=Project 1B Project;Persist Security Info=True;User Id=admin012;Password=Password012#;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;"
   },
   "Logging": {
     "LogLevel": {
